backend/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
class-registry==2.1.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
defusedxml==0.7.1
Django==3.2.18
django-cors-headers==3.14.0
django-extra-fields==3.0.2
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.4.1
drf-yasg==1.20.0
filters==1.3.2
flake8==5.0.4
gunicorn==20.0.4
idna==3.4
importlib-metadata==1.7.0
inflection==0.5.1
isort==5.11.5
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.0
Pillow==9.4.0
psycopg2-binary==2.9.5
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2022.7.1
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.0
sqlparse==0.4.3
typing-extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
zipp==3.15.0
